+++ socialmedia/src/main/java/br/com/socialmedia/socialmedia/dto/FollowersCountDto.java
@@ -3,7 +3,7 @@ package br.com.socialmedia.socialmedia.dto;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record FollowersCountDto(
-        @JsonProperty("user_id") Integer userId,
+        @JsonProperty("user_id") long userId,
         @JsonProperty("user_name") String userName,
-        @JsonProperty("followers_count") Integer followersCount
+        @JsonProperty("followers_count") long followersCount
 ) { }
