+++ socialmedia/src/main/java/br/com/socialmedia/socialmedia/dto/FollowersListDto.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.List;
 
 public record FollowersListDto(
-        @JsonProperty("user_id") Integer userId,
+        @JsonProperty("user_id") long userId,
         @JsonProperty("user_name") String userName,
         List<FollowDto> followers
 ) { }
