+++ socialmedia/src/main/java/br/com/socialmedia/socialmedia/dto/response/PromoCountResponse.java
@@ -1,10 +1,9 @@
 package br.com.socialmedia.socialmedia.dto.response;
 
-
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record PromoCountResponse(
-        @JsonProperty("user_id") int userId,
+        @JsonProperty("user_id") long userId,
         @JsonProperty("user_name") String userName,
-        @JsonProperty("promo_products_count") int promoProductsCount
+        @JsonProperty("promo_products_count") long promoProductsCount
 ) {}
