+++ socialmedia/src/main/java/br/com/socialmedia/socialmedia/dto/response/PromoPostsResponse.java
@@ -1,12 +1,11 @@
 package br.com.socialmedia.socialmedia.dto.response;
 
-
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import java.util.List;
 
 public record PromoPostsResponse(
-        @JsonProperty("user_id") int userId,
+        @JsonProperty("user_id") long userId,
         @JsonProperty("user_name") String userName,
         List<PostResponse> posts
 ) {}
